fix(ci): Fix env variables
