+++ .github/workflows/ci.yml
@@ -7,8 +7,8 @@ on:
       - 'main'
 
 env:
-  APP_GRAPHQL_CODEGEN_ENDPOINT: './backend/schema.graphql'
-  APP_GRAPHQL_ENDPOINT: 'http://web:8000/graphiql/'
+  NEXT_PUBLIC_GRAPHQL_CODEGEN_ENDPOINT: './backend/schema.graphql'
+  NEXT_PUBLIC_GRAPHQL_ENDPOINT: 'http://web:8000/graphql/'
   GITHUB_WORKFLOW: true
 
 jobs:
