+++ package.json
@@ -13,7 +13,7 @@
         "_prestart": "pnpm generate:type",
         "start": "next start",
         "typecheck": "tsc",
-        "prelint": "pnpm generate:type && pnpm typecheck",
+        "prelint": "pnpm generate:type",
         "lint": "next lint"
     },
     "engines": {
