--- tsconfig.scripts.json
@@ -1,9 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "noEmit": false
-    },
-    "include": ["scripts/fetchData.ts"]
-}
